+++ CHANGELOG.md
@@ -5,7 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.4] - 2018-11-28
+## [1.0.6] - 2018-11-29
+
+### Added
+
+- `esdoc` dev dependency as well as configuration file
+- Perfectly documented `Alpha` and `Attraction` behaviour classes
+- `docs/api` directory
+- `behaviours/constants.js` for defaults
+
+### Changed
+
+- `Alpha` behaviour now has `alphaA` and `alphaB` props rather than just `a` and `b`, unit tests updated accordingly
+
+### Fixed
+
+- There was a bug in the `Attraction` constructor that was causing `force` to be set to `NaN`, this has been fixed
+
+## [1.0.5] - 2018-11-28
 
 ### Fixed